cleaner

+++ test_status.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-export JENOVA_ROOT="/home/orpheus497/Documents/Projects/jenova"
+export JENOVA_ROOT="${JENOVA_ROOT:-$(cd "$(dirname "$0")" && pwd)}"
 export PID_FILE="/tmp/test_jenova.pid"
 export LLAMA_PORT=8080
 export JENOVA_CONNECT_HOST="127.0.0.1"
@@ -18,7 +18,7 @@ time $JENOVA_ROOT/bin/jenova-ca status
 
 # Case 3: Hung (listen but don't respond, forcing curl to timeout)
 echo -e "\nHung state:"
-nc -l -p 8080 &
+nc -l 8080 &
 NC_PID=$!
 time $JENOVA_ROOT/bin/jenova-ca status
-kill $NC_PID
+if kill -0 $NC_PID 2>/dev/null; then kill $NC_PID; fi
